Return value of sccp_rtp_get_sampleRate, is larger than uint8_t, changed to returning int
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@5362 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ src/sccp_rtp.c
@@ -273,7 +273,7 @@ uint8_t sccp_rtp_get_payloadType(const struct sccp_rtp * rtp, skinny_codec_t cod
 /*!
  * \brief Get Sample Rate
  */
-uint8_t sccp_rtp_get_sampleRate(skinny_codec_t codec)
+int sccp_rtp_get_sampleRate(skinny_codec_t codec)
 {
 	if (PBX(rtp_get_sampleRate)) {
 		return PBX(rtp_get_sampleRate) (codec);
